+++ src/features/seasoning/utils/nameValidationMessage.ts
@@ -2,7 +2,7 @@ import type { NameValidationError } from "@/utils/nameValidation";
 import {
   VALIDATION_CONSTANTS,
   VALIDATION_ERROR_MESSAGES,
-} from "@/features/seasoning/../../constants/validation";
+} from "@/constants/validation";
 
 /**
  * 調味料名バリデーションエラーのメッセージ表示
