+++ src/features/seasoning/utils/typeValidationMessage.ts
@@ -1,4 +1,4 @@
-import type { TypeValidationError } from "@/features/seasoning/../../utils/typeValidation";
+import type { TypeValidationError } from "@/utils/typeValidation";
 
 /**
  * タイプバリデーションエラーのメッセージ表示
