+++ src/hooks/useSeasoningSubmit.test.ts
@@ -31,7 +31,7 @@ const createMockSeasoningTypeInput = (
 });
 
 // モック作成
-vi.mock("../utils/imageValidation", () => ({
+vi.mock("@/utils/imageValidation", () => ({
   validateImage: vi.fn(),
 }));
 
@@ -179,7 +179,7 @@ describe("useSeasoningSubmit", () => {
 
   describe("リファクタリング後の単一責任テスト", () => {
     test("バリデーション処理が独立して動作すること", async () => {
-      const { validateImage } = await import("../utils/imageValidation");
+      const { validateImage } = await import("@/utils/imageValidation");
 
       const mockValidateImage = vi.mocked(validateImage);
 
@@ -211,7 +211,7 @@ describe("useSeasoningSubmit", () => {
     });
 
     test("フォーム状態リセット機能が独立して動作すること", async () => {
-      const { validateImage } = await import("../utils/imageValidation");
+      const { validateImage } = await import("@/utils/imageValidation");
 
       const mockValidateImage = vi.mocked(validateImage);
 
@@ -262,7 +262,7 @@ describe("useSeasoningSubmit", () => {
     });
 
     test("エラーメッセージが定数から取得されること", async () => {
-      const { validateImage } = await import("../utils/imageValidation");
+      const { validateImage } = await import("@/utils/imageValidation");
 
       const mockValidateImage = vi.mocked(validateImage);
 
@@ -301,7 +301,7 @@ describe("useSeasoningSubmit", () => {
     });
 
     test("バリデーション関数が適切な引数で呼び出されること", async () => {
-      const { validateImage } = await import("../utils/imageValidation");
+      const { validateImage } = await import("@/utils/imageValidation");
 
       const mockValidateImage = vi.mocked(validateImage);
 
@@ -344,7 +344,7 @@ describe("useSeasoningSubmit", () => {
       const mockOnSubmit = vi.fn().mockResolvedValue(undefined);
       const mockOnReset = vi.fn();
 
-      const { validateImage } = await import("../utils/imageValidation");
+      const { validateImage } = await import("@/utils/imageValidation");
 
       const mockValidateImage = vi.mocked(validateImage);
 
@@ -387,7 +387,7 @@ describe("useSeasoningSubmit", () => {
         .fn()
         .mockRejectedValue(new Error("ネットワークエラー"));
 
-      const { validateImage } = await import("../utils/imageValidation");
+      const { validateImage } = await import("@/utils/imageValidation");
 
       const mockValidateImage = vi.mocked(validateImage);
